+++ CHANGELOG.md
@@ -13,13 +13,14 @@
 - **lsp:** disabled open float in favor of lsp_lines
 - **lsp_lines:** changed packer declaration using as
 - **lsp_lines:** added lsp_lines
-- **pretty-fold:** set fold next max to 1
 - **pretty-fold:** added pretty fold
+- **pretty-fold:** set fold next max to 1
 - **scrollbar:** added scrollbar
 - **settings:** H now moved to close fold, L to open fold
 - **telescope:** added telescope env
 - **treesitter:** foldexprt now to treesitter
 - **which-key:** added dofile on <leader>pR
+- **windline:** added lsp signature
 
 ### Fix
 - **tabout:** fix whichkey missing '>'